feat: Add raycaster and mouse click event listener to create blocks at clicked position
The code changes add a raycaster and a mouse click event listener to the app.js file. This allows the user to create blocks at the position where they click on the screen. The raycaster is used to calculate the intersection between the mouse click and the ground mesh, and if there is an intersection, a block is created at that position. This feature enhances the interactivity of the application.