feat: Add OBJLoader for loading .obj files and createBlock function to load and position 3D objects
The code changes introduce the use of OBJLoader from the Three.js library to load .obj files. The createBlock function is updated to use the OBJLoader to load a specific .obj file and position it in the scene. This allows for the creation of 3D objects at the clicked position, enhancing the visual variety and realism of the application.